% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_seurat.R
\name{AddClusterIndicators}
\alias{AddClusterIndicators}
\title{Create indicator vectors for a categorical variable.}
\usage{
AddClusterIndicators(dge, results_path = NULL, ident.use = "ident",
  ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{If not null, save plots here.}

\item{ident.use}{Categorical vector to create indicators for}
}
\description{
Create indicator vectors for a categorical variable.
}
