% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{AddTotalHSP}
\alias{AddTotalHSP}
\title{Add up heat shock protein and chaperonin transcripts cell by cell.}
\usage{
AddTotalHSP(dge)
}
\description{
Add up heat shock protein and chaperonin transcripts cell by cell.
}
