% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annot_ident_plot}
\alias{annot_ident_plot}
\title{Annotate a plot with cluster labels}
\usage{
annot_ident_plot(dge, results_path = NULL, ident.use = "ident",
  figname = ident.use, height = 20, width = 20, axes = c("tSNE_1",
  "tSNE_2"), ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to put the resulting plot (directory). If NULL, no plot is saved.}

\item{ident.use}{What clusters to plot.}

\item{figname}{Name of the resulting file.}

\item{height}{@param width Passed to ggsave.}

\item{axes}{@param ... Passed to custom_feature_plot. Axes are also used to position labels directly over clusters.}
}
\description{
Annotate a plot with cluster labels
}
