% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_seurat.R
\name{AvailableData}
\alias{AvailableData}
\title{Get all available variable names (genes, identity classes, PCA embeddings, etc)}
\usage{
AvailableData(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a character vector of all eligible inputs to the `vars.all` argument of `FetchData`.
}
\description{
Get all available variable names (genes, identity classes, PCA embeddings, etc)
}
