% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{ClassifyMLR}
\alias{ClassifyMLR}
\title{Apply a machine learning classifier (from `train_save_classifier`) to new data}
\usage{
ClassifyMLR(dge, mlr_mod, return_type = "seurat", s = "lambda.min",
  do_knn_smooth = T, k = 30, dims.use = 30, reduction.use = "PC")
}
\arguments{
\item{dge}{a Seurat object}

\item{mlr_mod}{a glmnet multiclass logistic regression model.
You can feed the output of `train_save_classifier` into the `mlr_mod` argument.}

\item{return_type}{If "seurat" (default), returns a Seurat object with results in the metadata.}

\item{s}{Controls the regularization used for predictions.}

\item{do_knn_smooth}{Smooth test data via knn before predicting?}

\item{k}{Number of cells to average in knn smoothing}

\item{reduction.use}{@param dims.use For computation of KNN's in a low-dimensional space.}
}
\description{
Apply a machine learning classifier (from `train_save_classifier`) to new data
}
