% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{ClusterSmoothedGenes}
\alias{ClusterSmoothedGenes}
\title{Cluster genes by behavior over pseudotime.}
\usage{
ClusterSmoothedGenes(dge, results_path, smoothers, pt.use = "pseudotime",
  abcissae_kmeans = 20, num_clusters = NULL)
}
\arguments{
\item{dge}{should be a seurat object with a field "pseudotime". 
The field `dge@data` is accessed for expression levels -- for Eric's objects, the units will be log2(1+CP10K).}

\item{smoothers}{Output from SmoothGenes.}

\item{abcissae_kmeans}{Gene expression is fed into k-means as a series of predictions at successive time points.
The arguments says how many time points to predict and feed in (if length one) or what time points (if longer).}

\item{num_clusters}{Genes are partitioned into this many modules. If NULL (default) the value is selected via the gap statistics and their SEs using the method in the original gap statistic paper:

Tibshirani, R., Walther, G. and Hastie, T. (2001). Estimating the number 
of data clusters via the Gap statistic. Journal of the Royal Statistical Society B, 63, 411–423.
                                
There's one adjustment: this function will never use just one cluster. It will issue a warning and use 2.}
}
\value{
A list with elements:
\itemize{
\item cluster_mod: output from stats::kmeans 
\item abcissae_kmeans: pseudotime values at which smoothers were evaluated to form kmeans_features
\item kmeans_features: input for stats::kmeans 
\item gap_stats: output from cluster::clusGap
}
}
\description{
Cluster genes by behavior over pseudotime.
}
