% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{custom_feature_plot}
\alias{custom_feature_plot}
\title{Make tSNE plots (or PCA, or Monocle; it's customizable)}
\usage{
custom_feature_plot(dge, colour = NULL, subset_id = NULL,
  axes = c("tSNE_1", "tSNE_2"), fix_coords = T, show_axes = F,
  alpha = 1, size.use = NULL, cols.use = c("khaki1", "red"),
  use_rank = F, overplot_adjust = F, mode = "regular",
  num_bins = 50, contour_pct_outside = 5, contour_params = list(col =
  "grey", linetype = "dashed"), ...)
}
\arguments{
\item{dge}{Seurat object}

\item{colour}{Variable to fetch and plot. Length-1 character. If omitted, a dark grey plot is returned.}

\item{subset_id}{Vector of subsets to include on the plot. Cells not included will still be included via
geom_blank so as to preserve the scale of the plot.}

\item{axes}{Variables to fetch for use on X and Y axes. Defaults to tSNE embedding.}

\item{fix_coords}{Force the plot to be square? Defaults to T, because a rotated or reflected version of 
the tSNE embedding should convey the exact same relationships.}

\item{alpha}{Transparency. Number between 0 and 1. size.use}

\item{size.use}{Point size. If NULL, uses sensible defaults.}

\item{cols.use}{Colorscale. Character vector, hexadecimal or e.g. c("khaki1", "red"). 
For discrete data fetched for "colour", should be named with the variable's levels.}

\item{use_rank}{Transform expression level to ranks before plotting?}

\item{overplot_adjust}{Deprecated. Use mode arg instead.}

\item{mode}{Different styles of plot. Should be one of \code{c("regular", "points", "overplot_adjust", "umi_density", "umidensity")}. Case insensitive, so "ReGuLaR" works.}

\item{num_bins}{For hexagonal binning when mode='overplot_adjust'. Number of bins across the X axis.}

\item{contour_pct_outside}{Used in UMI density mode. This number (an integer in [1, 100]) specifies
the tightness of the contour drawn on the plot -- how many cells are allowed to fall outside it?}

\item{contour_params}{Used in UMI density mode. This list of parameters is
passed to geom_path for drawing the outline around the cells.}

\item{...}{Passed into functions that preprocess before plotting. Currently only used 
in mode 'umidensity' as extra params for the 2D KDE function 'ggtern::kde2d.weighted'.}
}
\description{
Make tSNE plots (or PCA, or Monocle; it's customizable)
}
