% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{DiffExpByType}
\alias{DiffExpByType}
\title{Test for effects of a treatment within cell types.}
\usage{
DiffExpByType(dge, treatment_var, treatment_1, treatment_2,
  cell_type_var = "classifier_ident", ...)
}
\arguments{
\item{dge}{Seurat object}

\item{treatment_var}{Passed to FetchData.}

\item{treatment_1}{@param treatment_2 Levels of treatment to compare.}

\item{cell_type_var}{Passed to FetchData.}
}
\description{
Test for effects of a treatment within cell types.
}
