% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{display_fractional_identity}
\alias{display_fractional_identity}
\title{Rank cells by fractional identity and display them with a KDE.}
\usage{
display_fractional_identity(dge_test, dge_train, main, ident.use,
  genes.use = NULL, test.ident.use = NULL, test.col = NULL)
}
\arguments{
\item{dge_test}{@param dge_train Test and training sets (seurat objects).}

\item{main}{Plot title}

\item{ident.use}{Field to extract as training labels.}

\item{genes.use}{Genes to train on.}

\item{test.ident.use}{@param test.col For plotting the distribution of predictions by group.}
}
\description{
Rank cells by fractional identity and display them with a KDE.
}
