% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{DisplayTrainingResults}
\alias{DisplayTrainingResults}
\title{Summarize results of training a multinomial logistic regression model to classify scRNA data.}
\usage{
DisplayTrainingResults(dge_train, results_path, ident.use, mlr_mod, ...,
  plotting_args = NULL)
}
\arguments{
\item{dge_train}{The Seurat object used to train the MLR model.}

\item{results_path}{Where to put plots.}

\item{ident.use}{The field used as the training set labels.}

\item{mlr_mod}{A multinomial logistic regression model from the glmnet package}

\item{...}{Extra params passed to ClassifyMLR}

\item{plotting_args}{Plotting params passed to save_feature_plots}
}
\description{
Summarize results of training a multinomial logistic regression model to classify scRNA data.
}
