% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapping.R
\name{DoHeatmapFast}
\alias{DoHeatmapFast}
\title{Save a big PDF file to `<results_path>/<main>.pdf` containing a heatmap of gene expression levels.}
\usage{
DoHeatmapFast(dge, results_path, genes.use = dge@var.genes,
  genes.preview = c(), genes.label = c(), main = "heatmap_cellwise",
  ident.use = "ident", norm_fun = div_by_max, cluster_colors = NA,
  cluster_order = NULL, num_pc = 20,
  col = colorRampPalette(c("khaki1", "red"))(20), width = 10,
  height = 10, ...)
}
\arguments{
\item{dge}{a Seurat object}

\item{results_path}{should be a character such that `dir.exists( results_path )`}

\item{genes.use}{Genes to include in the heatmap. Character vector.}

\item{genes.preview}{If this has length > 0, function will produce a quick preview containing these genes.}

\item{main}{Figure title and name of saved file.}

\item{ident.use}{Used to set up the ColSideColors in heatmap.2.}

\item{norm_fun}{Applied to each gene.}

\item{cluster_colors}{List of colors. Should be named with levels of the FetchData(dge, ident.use).
Used to set up the ColSideColors in heatmap.2. If you don't want any, put NULL. 
If NA, defaults to a sensible value.}

\item{cluster_order}{What order do you want the clusters in?}

\item{num_pc}{Number of PC's to use when reordering cells}

\item{col}{a vector of color names for use in the main body of the heatmap.}

\item{width}{@param height Dimensions of saved plot in inches.}

\item{...}{Passed on to heatmap.2.}
}
\description{
Save a big PDF file to `<results_path>/<main>.pdf` containing a heatmap of gene expression levels.
}
\details{
Each column is a cell and each row is a gene. Each gene is rescaled so that its peak expression is 1.
This facilitates comparison within genes and across cells, though it's bad for comparison across genes.
}
