% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{explore_embeddings}
\alias{explore_embeddings}
\title{Quickly explore many parameter settings}
\usage{
explore_embeddings(dge, results_path, all_params,
  whitelist = AvailableData(dge), blacklist = c(), regress_out = c(),
  plot_each_iteration = "orig.ident", do_cca = F,
  align_embeddings = F, group.by = NULL, group1 = NULL,
  group2 = NULL, ...)
}
\arguments{
\item{dge}{Seurat object.}

\item{results_path}{}

\item{all_params}{Dataframe specifying multiple runs of the pipeline. Must have these columns: 
\itemize{
\item "latent_dimension": integer. Latent dimension to use in PCA or CCA.
\item "clust_method": Character. Each entry should be a valid input for the 'method' arg of 'cluster_wrapper'. (Currently 'SNN' or 'DBSCAN').
\item "clust_granularities_as_string": Character. Comma-separated list of numbers to be used for G.use (if DBSCAN) or res (if SNN). Allows multiple clusterings based on the same PC list. 
\item "excess_var_cutoff": Numeric. Params for variable gene selection. Units are in multiples of the local median coefficient of variation, i.e. 0.8 means a gene is included if it exceeds 80% of the local median of the CV, binning based on expression level.
\item "log_expr_cutoff": Numeric. Genes below this expression level are excluded. Done by averaging the '@data' slot (so, log1p-scale expression).
}}

\item{whitelist}{Everything not on this list is excluded.}

\item{blacklist}{Genes to exclude. Typically nothing (default), but could be cell cycle genes (try}

\item{regress_out}{Covariates to regress out via Seurat::RegressOut. . 
\code{ thymusatlastools2::get_macosko_cc_genes() \%>\% Reduce(f=union) } or ribosomal 
subunit genes. Make sure you get the species right!}

\item{plot_each_iteration}{A list of featureplots to show at each iteration.}

\item{do_cca}{@param align_embeddings If do_cca, do CCA instead of PCA. If align_embeddings, align results using DTW.}

\item{group.by}{@param group1 @param group2 @param ... Passed to RunCCA.}
}
\description{
Quickly explore many parameter settings
}
