% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_seurat.R
\name{extract_tip_order}
\alias{extract_tip_order}
\title{Get the order of the tips from an object of class phylo from the APE package.}
\usage{
extract_tip_order(ape_phylo_object)
}
\description{
Get the order of the tips from an object of class phylo from the APE package.
}
