% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{FacetPlotGeneClusters}
\alias{FacetPlotGeneClusters}
\title{Draw overlaid line plots of gene clusters from output of `ClusterSmoothedGenes`.}
\usage{
FacetPlotGeneClusters(dge, results_path, pt.use = "pseudotime",
  cluster_mod, kmeans_features, abcissae_kmeans, facet_ncol = NULL)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to save plot}

\item{pt.use}{What to extract as the pseudotime field.}

\item{cluster_mod}{Kmeans output with genewise cluster assignments}

\item{abcissae_kmeans}{Pseudotime values where gene smoothers were evaluated}

\item{facet_ncol}{For resulting facet plots, number of columns.}
}
\description{
Draw overlaid line plots of gene clusters from output of `ClusterSmoothedGenes`.
}
