% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{FindHighlySpecific}
\alias{FindHighlySpecific}
\title{Find genes highly specific to a single cluster by comparing to the next-highest (or third-highest, or r-th highest) cluster.}
\usage{
FindHighlySpecific(dge, ident.use = "ident", n = 10, r = 2,
  AGG_FUN = prop_nz, genes.use = dge@data \%>\% rownames)
}
\arguments{
\item{dge}{}

\item{ident.use}{A categorical variable obtainable via FetchData. Clusters to find markers for.}

\item{n}{How many genes to return for each cluster.}

\item{r}{Each cluster is compared to the second highest cluster if r=2, third if r=3, et cetera. The default is r=2, but if you have (for example) two FOXN1+ clusters near the tip of a differentiation trajectory, you could set it a little higher to still rank FOXN1 highly.}

\item{AGG_FUN}{Given a numeric vector x of cell values for one cluster, 
aggregate them into a value to be used for ranking clusters. Defaults to proportion of nonzeroes.}

\item{genes.use}{Genes to start from.

For each gene, rank clusters (e.g. by proportion of cells expressing that gene). 
Find the peak cluster. 
Then subtract the second (rth) cluster's expression from the peak cluster expression. 
In each cluster, take all genes peaking in that cluster, rank them by the result of the 
previous calculation, and return the top n.}
}
\description{
Find genes highly specific to a single cluster by comparing to the next-highest (or third-highest, or r-th highest) cluster.
}
