% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{FindMarkersFlex}
\alias{FindMarkersFlex}
\title{Test for markers flexibly from a Seurat object.}
\usage{
FindMarkersFlex(object, genes.use, ident.use, ident.1, ident.2 = object
  \%>\% FetchData(ident.use) \%>\% extract2(1) \%>\% unique \%>\%
  setdiff(ident.1), order_by_var = "avg_logFC", ...)
}
\arguments{
\item{genes.use}{@param ... Passed into FindMarkers. genes.use is filtered to exclude unavailable genes.}

\item{ident.use}{Fetched via FetchData to define the groups being tested.}

\item{ident.1}{@param ident.2 Levels of ident.use to compare. Character vectors expected. These may have length > 1.}

\item{order_by_var}{Column to sort the results by.}
}
\description{
Convenience wrapper for Seurat::FindMarkers.
}
