% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{get_active_genes}
\alias{get_active_genes}
\title{Given a test set, training set, and classifier, find genes that behave the same in test and train sets.}
\usage{
get_active_genes(dge_test, dge_train, train_ident, mlr_mod, ...)
}
\arguments{
\item{dge_test}{Seurat object}

\item{dge_train}{Seurat object}

\item{train_ident}{Metadatum from dge_train used as training labels.}

\item{mlr_mod}{multiclass logistic regression output from glmnet::cv.glmnet}

\item{...}{Additional args passed to get_genes_by_cluster}
}
\description{
Given a test set, training set, and classifier, find genes that behave the same in test and train sets.
}
