% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_cell_surface}
\alias{get_cell_surface}
\title{Return cell surface molecules from Panther}
\usage{
get_cell_surface(species)
}
\arguments{
\item{species}{'human' or 'mouse'.
Currently uses categories CellAdhesion, CellJunction, ExtracellMatrix, Immunity, Receptor, and SignalingMol.}
}
\description{
Return cell surface molecules from Panther
}
