% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{get_frac_ident_pure}
\alias{get_frac_ident_pure}
\title{Compute fractional identities for a single cell.}
\usage{
get_frac_ident_pure(query_cell, pseudo_bulk)
}
\description{
Compute fractional identities for a single cell.
}
