% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{get_genes_by_cluster}
\alias{get_genes_by_cluster}
\title{Extract active features for a given class from a glmnet multiclass logistic regression model.}
\usage{
get_genes_by_cluster(mlr_mod, cluster_name, ...)
}
\description{
Extract active features for a given class from a glmnet multiclass logistic regression model.
}
