% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{get_maehrlab_data}
\alias{get_maehrlab_data}
\title{Alias for load_maehrlab_data}
\usage{
get_maehrlab_data(sample_ids = "all", test_mode = F, lincRNA = F)
}
\description{
Alias for load_maehrlab_data
}
