% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{get_metadata_path}
\alias{get_metadata_path}
\title{Get path to thymus atlas sample metadata (and data).}
\usage{
get_metadata_path()
}
\description{
Get path to thymus atlas sample metadata (and data).
}
