% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{GetDimMarchenkoPastur}
\alias{GetDimMarchenkoPastur}
\title{Estimate which PC's to keep using the Marchenko-Pastur asymptotic upper bound.}
\usage{
GetDimMarchenkoPastur(dge)
}
\arguments{
\item{dge}{A Seurat object with gene selection and PCA done.}
}
\description{
Estimate which PC's to keep using the Marchenko-Pastur asymptotic upper bound.
}
