% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{GetDynamicGenes}
\alias{GetDynamicGenes}
\title{Filter genes based on qvals from ANOVA of constant model (null) versus piecewise linear model (alternative)}
\usage{
GetDynamicGenes(dge, num_periods_initial_screen,
  genes.use = rownames(dge@data), pt.use = "pseudotime")
}
\arguments{
\item{dge}{@param pt.use Seurat object and name of pseudotime field to select from it.}

\item{num_periods_initial_screen}{Cells are partitioned into this many pseudotime periods (equal number of cells in each).}
}
\description{
Filter genes based on qvals from ANOVA of constant model (null) versus piecewise linear model (alternative)
}
