% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapping.R
\name{gg_add_colorbar}
\alias{gg_add_colorbar}
\title{Add a colourbar to a ggplot object, fixing the fill to allow a separate colorscale.}
\usage{
gg_add_colorbar(plot, x, my_labels, col = NULL, width_ = 0.1,
  position = -0.1, is_horiz = T)
}
\description{
Add a colourbar to a ggplot object, fixing the fill to allow a separate colorscale.
}
