% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{HeatmapGeneClusters}
\alias{HeatmapGeneClusters}
\title{Draw heatmaps of gene clusters from output of `ClusterSmoothedGenes`.}
\usage{
HeatmapGeneClusters(dge, results_path, pt.use = "pseudotime",
  branch.use = NULL, cluster_mod, smoothers, gap_size = NULL,
  genes_use = NULL, genes_to_label = NULL, cols.use = blue_gray_red,
  branch_colors = c((scales::hue_pal())(2), "purple", "gray"))
}
\arguments{
\item{dge}{Seurat object with raw data and pseudotime metadata used to train smoothers. 
If metadata `simple_branch` is present, function is hardwired to look for "mTEC", "cTEC", "branchpoint", and "progenitor"
and make a heatmap similar to figure 2 in http://dx.doi.org/10.1101/122531.}

\item{results_path}{Where to save plots and files.}

\item{pt.use}{What to extract as the pseudotime field.}

\item{branch.use}{For branching heatmaps, field to extract and use as branch ID. If null, regular heatmap shown.}

\item{cluster_mod}{K-means output with cluster labels for the genes in `smoothers` and also with cluster centers.}

\item{smoothers}{List of regression models for the genes in `cluster_mod` and `gene_stats`.}

\item{gap_size}{White bars separating clusters are formed by adding fake genes. gap_size is how many fake genes per bar.}

\item{genes_use}{Gene names or anything in `AvailableData(dge)`.}

\item{genes_to_label}{Subset of `genes_use` to write tick labels for.}

\item{cols.use}{@param branch_colors Colors to use for expression and for cell clusters}
}
\description{
Draw heatmaps of gene clusters from output of `ClusterSmoothedGenes`.
}
