% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{InstantOverview}
\alias{InstantOverview}
\title{Create an information-dense metadata variable: clusters annotated with marker genes.}
\usage{
InstantOverview(dge, results_path = NULL, genes.use,
  gene_list_name = deparse(substitute(genes.use)), n = 5,
  ident.use = "ident", include_size = T, ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to put the resulting plot. If NULL, no plot is saved.}

\item{genes.use}{}

\item{gene_list_name}{Goes into name of resulting plot.}

\item{n}{Number of genes to label each cluster with}

\item{ident.use}{Clusters to annotate}

\item{...}{Additional parameters for annot_ident_plot}
}
\description{
Create an information-dense metadata variable: clusters annotated with marker genes.
}
