% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{intersect_preserving_order}
\alias{intersect_preserving_order}
\title{Like \code{base::intersect}, but preserves order of first arg.}
\usage{
intersect_preserving_order(x, y)
}
\description{
Like \code{base::intersect}, but preserves order of first arg.
}
