% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{LabelDoublets}
\alias{LabelDoublets}
\title{Label doublets using the DoubletFinder package.}
\usage{
LabelDoublets(dge, rep_field = "orig.ident", results_path = NULL)
}
\arguments{
\item{dge}{seurat object.}

\item{results_path}{Where to save a brief table summarizing results. If null, nothing gets saved.}

\item{orig.ident}{Replicate. Only looks for doublets within reps.}
}
\description{
Label doublets using the DoubletFinder package.
}
