% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{load_thymus_profiling_data}
\alias{load_thymus_profiling_data}
\title{Alias for load_maehrlab_data}
\usage{
load_thymus_profiling_data(sample_ids = "all", test_mode = F,
  lincRNA = F)
}
\description{
Alias for load_maehrlab_data
}
