% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{master_pt}
\alias{master_pt}
\title{Run Monocle, DPT, or PCA-based pseudotime analysis on a Seurat object.}
\usage{
master_pt(dge, results_path, method = "monocle", earliest_day = NULL,
  return_type = c("seurat", "all"), reset_var_genes = T,
  mp = list(reset_var_genes = reset_var_genes, log_scale_expr_thresh =
  0.1, excess_disp = 1, num_mature_types = NULL, reduction_method =
  "DDRTree"), ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to save tables and plots}

\item{earliest_day}{Numeric. Used to root the Monocle trajectory; gets compared to `FetchData(dge, "eday")[[1]]` via `==`.}

\item{return_type}{'seurat' for the original object with fields added. 'all' for a list that also contains the monocle or dpt object.}

\item{mp}{Passed to call_monocle_on_seurat as monocle_params arg.}

\item{...}{Additional parameters passed to call_monocle_on_seurat.}
}
\description{
Run Monocle, DPT, or PCA-based pseudotime analysis on a Seurat object.
}
