% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_dupe_rows}
\alias{merge_dupe_rows}
\title{Merge duplicate rows of a Matrix (from the Matrix package).}
\usage{
merge_dupe_rows(X, row.names, FUN = Matrix::colSums, verbose = F)
}
\arguments{
\item{X}{Matrix to start with.}

\item{FUN}{Merging function.}
}
\description{
Merge duplicate rows of a Matrix (from the Matrix package).
}
