% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapping.R
\name{OrderCellsWithinClusters}
\alias{OrderCellsWithinClusters}
\title{Return a cell ordering where predefined clusters remain contiguous.}
\usage{
OrderCellsWithinClusters(dge, ident.use, coords.use = paste0("PC", 1:30),
  cluster_order = NULL)
}
\description{
Return a cell ordering where predefined clusters remain contiguous.
}
