% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R, R/heatmapping.R
\name{OrderRowsForHeatmap}
\alias{OrderRowsForHeatmap}
\title{Given a column ordering, produce a pleasing row ordering.}
\usage{
OrderRowsForHeatmap(X, outlier_cutoff = 0.2, wrap = T,
  REORDERFUN_outlier = function(x, ...) {     x \%>\% dist \%>\% hclust
  \%>\% as.dendrogram \%>\% (stats::reorder) }, ...)

OrderRowsForHeatmap(X, outlier_cutoff = 0.2, wrap = T,
  REORDERFUN_outlier = function(x, ...) {     x \%>\% dist \%>\% hclust
  \%>\% as.dendrogram \%>\% (stats::reorder) }, ...)
}
\arguments{
\item{X}{Set of data to order. Coerced to dataframe.}

\item{outlier_cutoff}{Numeric vector of length 1. See details.}

\item{wrap}{Logical. Treat column n as a neighbor of column 1?}

\item{REORDERFUN_outlier}{Function accepting a df X and returning a vector of indices. 
See details.}

\item{...}{Extra args passed to REORDERFUN_outlier.}

\item{X}{Set of data to order. Coerced to dataframe.}

\item{wrap}{Logical. Treat column n as a neighbor of column 1?}

\item{outlier_cutoff}{Numeric vector of length 1. See details.}

\item{REORDERFUN_outlier}{Function accepting a df X and returning a vector of indices. 
See details.}

\item{...}{Extra args passed to REORDERFUN_outlier.}
}
\value{
Returns a permutation of rownames(X).

This function attempts to order rows for a heatmap to match a given column ordering. 
It attempts to create diagonal structure and it pays attention to which columns are adjacent.
The algorithm first orders by where the max value occurs: if row Alice peaks in column 1 and 
row Bob peaks in column 2, then Alice will precede Bob in the final ordering.
If Bob and Cassandra both peak in column 2, then the tiebreaker is column 3 minus column 1.

Returns a permutation of rownames(X).

This function attempts to order rows for a heatmap to match a given column ordering. 
It attempts to create diagonal structure and it pays attention to which columns are adjacent.
The algorithm first orders by where the max value occurs: if row Alice peaks in column 1 and 
row Bob peaks in column 2, then Alice will precede Bob in the final ordering.
If Bob and Cassandra both peak in column 2, then the tiebreaker is column 3 minus column 1.
}
\description{
Given a column ordering, produce a pleasing row ordering.

Given a column ordering, produce a pleasing row ordering.
}
