% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{plot_classifier_coefs}
\alias{plot_classifier_coefs}
\title{Display coefficients from a glmnet multiclass logistic regression model.}
\usage{
plot_classifier_coefs(mlr_mod, ...)
}
\description{
Display coefficients from a glmnet multiclass logistic regression model.
}
