% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{Read10X}
\alias{Read10X}
\title{Import data in 10X (v2 or v3) format. Adapted from the eponymous Seurat function.}
\usage{
Read10X(data.dir = NULL, version = NULL)
}
\arguments{
\item{data.dir}{Where to look.}

\item{version}{Version of 10X software used.}

\item{object}{Seurat object.}
}
\description{
Import data in 10X (v2 or v3) format. Adapted from the eponymous Seurat function.
}
