% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_seurat.R
\name{RenameByDendro}
\alias{RenameByDendro}
\title{Re-number clusters to be compatible with dendrogram.}
\usage{
RenameByDendro(object)
}
\description{
Re-number clusters to be compatible with dendrogram.
}
