% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SavePlotGrid}
\alias{SavePlotGrid}
\title{Save a rectangular grid of featureplots.}
\usage{
SavePlotGrid(dge, results_path, gene_list, gene_list_name, ncol, width,
  height, memory_saver = T, use_raw = F, leg_pos = c(0.9, 0.9),
  title_pos = c(0.5, 0.9), edit_fun = (function(p) return(p)), ...)
}
\arguments{
\item{dge}{Seurat object.}

\item{results_path}{Where to put the PDF.}

\item{gene_list}{What genes (features) to include.}

\item{gene_list_name}{Becomes name of plot.}

\item{ncol}{How wide is the grid (in panels)?}

\item{width}{How wide is the grid (in inches)?}

\item{height}{How tall is the grid (inches)?}

\item{memory_saver}{Delete scale data of input to free memory and speed things up?}

\item{leg_pos, title_pos}{Element positions within the plot. Length 2 character vector with entries between 0 and 1.}

\item{edit_fun}{This function is applied to every plot to put on the finishing touches.}
}
\description{
Save a rectangular grid of featureplots.
}
