% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{seuratify}
\alias{seuratify}
\title{Make arrays into Seurat objects.}
\usage{
seuratify(raw_dge, results_path = NULL, min.genes = 0, min.cells = 0,
  ...)
}
\description{
Keeps all genes expressed in at least
(by default) 3 cells and keeps all cells with at least 1000 genes. It reports some 
summary figures, plotting number of genes by cell, num UMIs by cell, and number of cells by
gene.
}
