% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{seuratify_thy_data}
\alias{seuratify_thy_data}
\title{Alias for seuratify, for backwards compatibility.}
\usage{
seuratify_thy_data(raw_dge, results_path = NULL, min.genes = 0,
  min.cells = 0, ...)
}
\description{
Alias for seuratify, for backwards compatibility.
}
