% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_seurat.R
\name{SeuratMerge}
\alias{SeuratMerge}
\title{Merge two Seurat objects.}
\usage{
SeuratMerge(dge1, dge2, preserve_ident = F,
  vars.keep = intersect(names(dge1@meta.data), names(dge2@meta.data)))
}
\description{
By default, preserves any column of @meta.data shared between both objects.  
You can also specify what variables to keep. They will be added to meta.data in 
the output, warning and padding with zeroes if either object lacks any var in vars.keep.
}
