% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{SmoothGenes}
\alias{SmoothGenes}
\title{Characterize genes by behavior over pseudotime, returning cluster assignments and smoothed expression values.}
\usage{
SmoothGenes(dge, genes.use, pt.use = "pseudotime")
}
\arguments{
\item{dge}{should be a seurat object with a field "pseudotime". 
The field `dge@data` is accessed for expression levels -- for Eric's objects, the units will be log2(1+CP10K).}

\item{genes.use}{@param pt.use Fields to fetch for regression (y and x values respectively).}
}
\description{
Characterize genes by behavior over pseudotime, returning cluster assignments and smoothed expression values.
}
