% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_raw.R
\name{Write10X}
\alias{Write10X}
\title{Export data in 10X format.}
\usage{
Write10X(object, data.dir, metadata.include = colnames(object@meta.data),
  slot = "raw.data", include_all = FALSE, ...)
}
\arguments{
\item{object}{Seurat object.}

\item{data.dir}{Where to put output out to.}

\item{metadata.include}{Which columns to include from the meta.data slot. You can put anything that works in FetchData, so tSNE and PCA/ICA coordinates are OK too. These will go to an additional flat file called metadata.tsv .}

\item{slot}{Passed to GetAssayData.}

\item{include_all}{If FALSE (default), include only what you see on the tSNE (formally, what's in the cell.names slot). This option exists because what you see on the tSNE is sometimes out of sync with what's in the raw.data slot: the raw.data slot doesn't get subsetted during operations such as SubsetData.}

\item{...}{Passed to GetAssayData.}
}
\description{
Export data in 10X format.
}
